Revert "Merge pull request #42 from lindgrenj6/crawl_phase_azure"
This reverts commit cc9650da98a39c1368687117f741ff530bcfbae3, reversing
changes made to 25bc3d72730e9ea2d4ce291ebe8041d532d5bf10.
