+++ Dockerfile
@@ -1,4 +1,5 @@
-FROM registry.access.redhat.com/ubi8/ubi-minimal:8.4-210 as build
+FROM registry.access.redhat.com/ubi8/ubi-minimal:8.4-205 as build
+MAINTAINER jlindgre@redhat.com
 
 RUN mkdir /build
 WORKDIR /build
@@ -6,17 +7,17 @@ WORKDIR /build
 RUN microdnf install go
 
 COPY go.mod .
-RUN go mod download
+RUN go mod download 
 
 COPY . .
 RUN go build
 
-FROM registry.access.redhat.com/ubi8/ubi-minimal:8.4-210
+FROM registry.access.redhat.com/ubi8/ubi-minimal:8.4-205
 COPY --from=build /build/sources-superkey-worker /sources-superkey-worker
 
-# install az cli from micro$oft
-RUN rpm --import https://packages.microsoft.com/keys/microsoft.asc && \
-    echo -e "[azure-cli]\nname=Azure CLI\nbaseurl=https://packages.microsoft.com/yumrepos/azure-cli\nenabled=1\ngpgcheck=1\ngpgkey=https://packages.microsoft.com/keys/microsoft.asc" | tee /etc/yum.repos.d/azure-cli.repo && \
-    microdnf install azure-cli
+RUN curl -L -o /usr/bin/haberdasher \
+    https://github.com/RedHatInsights/haberdasher/releases/latest/download/haberdasher_linux_amd64 && \
+    chmod 755 /usr/bin/haberdasher
 
-ENTRYPOINT ["/sources-superkey-worker"]
+ENTRYPOINT ["/usr/bin/haberdasher"]
+CMD ["/sources-superkey-worker"]
