+++ Makefile
@@ -15,9 +15,6 @@ container:
 run: build
 	./sources-superkey-worker
 
-inlinerun:
-	go run *.go
-
 fancyrun: build
 	./sources-superkey-worker | grep '^{' | jq -r .
 
@@ -37,4 +34,4 @@ lint:
 gci:
 	golangci-lint run -E gci --fix
 
-.PHONY: build container run fancyrun runcontainer clean debug tidy debug remotedebug lint gci inlinerun
+.PHONY: build container run fancyrun runcontainer clean debug tidy debug remotedebug lint gci
