+++ provider/amazon_provider.go
@@ -1,6 +1,8 @@
 package provider
 
 import (
+	"crypto/rand"
+	"encoding/hex"
 	"encoding/json"
 	"fmt"
 	"path"
@@ -136,6 +138,17 @@ func (a *AmazonProvider) ForgeApplication(request *superkey.CreateRequest) (*sup
 	f.CreatePayload(&username, nil, &appType)
 
 	return f, nil
+}
+
+// generateGUID() generates a short guid for resources
+func generateGUID() (string, error) {
+	bytes := make([]byte, 8)
+	_, err := rand.Read(bytes)
+	if err != nil {
+		return "", err
+	}
+
+	return hex.EncodeToString(bytes), nil
 }
 
 // getShortName(string) generates a name off of the application type
