+++ provider/forge.go
@@ -13,8 +13,7 @@ import (
 func Forge(request *superkey.CreateRequest) (*superkey.ForgedApplication, error) {
 	client, err := getProvider(request)
 	if err != nil {
-		// the request is required to post back the resources for the resource name.
-		return &superkey.ForgedApplication{Request: request}, err
+		return nil, err
 	}
 	f, err := client.ForgeApplication(request)
 
@@ -58,6 +57,7 @@ func getProvider(request *superkey.CreateRequest) (superkey.Provider, error) {
 
 	switch request.Provider {
 	case "amazon":
+		// client, err := amazon.NewClient(os.Getenv("AWS_ACCESS"), os.Getenv("AWS_SECRET"), getStepNames(request.SuperKeySteps)...)
 		client, err := amazon.NewClient(
 			*auth.Username,
 			*auth.Password,
@@ -68,15 +68,6 @@ func getProvider(request *superkey.CreateRequest) (superkey.Provider, error) {
 		}
 
 		return &AmazonProvider{Client: client}, nil
-
-	case "azure":
-		if auth.Extra == nil || auth.Extra.Azure == nil || *auth.Extra.Azure.TenantId == "" {
-			l.Log.Errorf("superkey credential %v missing tenant id", request.SuperKey)
-			return nil, fmt.Errorf("superkey credential %v missing tenant id", request.SuperKey)
-		}
-
-		return &AzureProvider{Username: *auth.Username, Password: *auth.Password, Tenant: *auth.Extra.Azure.TenantId}, nil
-
 	default:
 		return nil, fmt.Errorf("unsupported auth provider %v", request.Provider)
 	}
